test(bpmn): seed the hidden matrix from pre_run like every other e2e
The 12 hidden cases were a 240-line table inside the grader with a nonce
generated at import time. Every business-outcome e2e in this repo instead
writes its cases to seed.json from a `pre_run` step and has the checker read
them (uipath-maestro-flow/e2e/escalation_slack_alert/seed.py and siblings),
so the matrix is data the task owns rather than code, and the per-run id is
generated once where the fixtures are.

- new `seed.py` owns `Scenario`, the `scenario()` factory, the 12 cases and
  `SEEDED_DUPLICATE_KEY`; writes seed.json with a `uuid4` run id
- `pre_run` runs it before the agent starts
- the grader gains `load_scenarios()`, called first thing in `main()` and
  failing loudly if seed.json is absent. Read at grade time, not import
  time, so the module stays importable in a bare workspace -- matching the
  flow checkers' shape.
- a silent built-in fallback would be worse than failing: a stale matrix
  graded against a fresh run id makes every correlation assertion compare
  the wrong values, and the run would look like a business failure
- seed.json is gitignored; the unit tests build one in-process from seed.py,
  which keeps the two in lockstep -- a seed-shape change the loader cannot
  parse now fails in tests rather than in a live run

Sizes: behavior checker 2097 -> 1868, seed.py 343 new.

Also fixed two things this surfaced: my own post_run conformance test counted
pre_run's command too (it globbed the whole file), and seed.py needed
registering in sys.modules before exec_module because it defines a
@dataclass, which dataclasses resolves through sys.modules.
