+++ .gitignore
@@ -106,3 +106,7 @@ package-lock.json
 # CWD: the ephemeral solution's .uipx (so post_run can glob it) plus
 # per-scenario debug logs. Run artifact, never committed.
 .customer-escalation-live/
+
+# Hidden scenario matrix generated by the escalation eval's pre_run seed step.
+# Regenerated per run with a fresh run id; never committed.
+tests/tasks/uipath-maestro-bpmn/interactive/customer_escalation_triage/seed.json
