+++ tests/tasks/uipath-maestro-bpmn/interactive/customer_escalation_triage/check_customer_escalation_behavior.py
@@ -21,7 +21,6 @@ import json
 import contextlib
 import os
 import re
-import secrets
 import subprocess
 import sys
 import time
@@ -74,12 +73,6 @@ EXPECTED_LIVE_TARGET = {
     "Organization": "codereval",
     "Tenant": "DefaultTenant",
 }
-RUN_NONCE = secrets.token_hex(6)
-# Every UpdateExisting scenario seeds a real Jira issue first, and main
-# overwrites duplicateIssueKey with that issue's key (padded, to prove the
-# process trims it). The table carries this sentinel so nobody reads a fake
-# key like "JIRA-42" as the value actually sent.
-SEEDED_DUPLICATE_KEY = "__SEEDED_JIRA_KEY_SET_AT_RUNTIME__"
 # Durable cleanup journal, relative to the sandbox CWD. The in-band leases
 # free resources from a `finally` block, which a SIGKILL on the graded command
 # skips entirely. Every created id is therefore appended here the moment it
@@ -189,293 +182,70 @@ class Scenario:
     uses_error_boundary: bool = False
 
 
-def scenario(
-    name: str,
-    *,
-    customer_tier: str = "Standard",
-    crm_matches: int = 1,
-    service_state: str = "Available",
-    workaround: bool = True,
-    duplicate_key: str = "",
-    attachments: tuple[str, ...] = (),
-    agent_valid: bool = True,
-    jira_available: bool = True,
-    auto_send: bool = False,
-    business_impact: str | None = None,
-    expected: dict[str, Any],
-    uses_error_boundary: bool = False,
-) -> Scenario:
-    correlation = f"EVAL-live-alpha-{RUN_NONCE}-{name}-Exact"
-    values = {
-        "customerTier": customer_tier,
-        "crmMatchCount": crm_matches,
-        "serviceState": service_state,
-        "workaroundAvailable": workaround,
-        "duplicateIssueKey": duplicate_key,
-        "attachments": [
-            {"name": item, "driveFileId": "__DRIVE_SOURCE_FILE_ID__"}
-            for item in attachments
-        ],
-        "agentOutputValid": agent_valid,
-        "jiraAvailable": jira_available,
-        "autoSendEnabled": auto_send,
-        "businessImpact": (
-            business_impact
-            if business_impact is not None
-            else f"Hidden Alpha scenario {name}"
-        ),
-        "correlationId": correlation,
-    }
-    complete_expected = dict(expected)
-    complete_expected["caseKey"] = correlation
-    return Scenario(
-        name=name,
-        inputs=values,
-        outputs=complete_expected,
-        attachment_iterations=attachments
-        if expected["attachmentAction"] == "SaveToDrive"
-        else (),
-        uses_error_boundary=uses_error_boundary,
-    )
+SEED_FILE = Path("seed.json")
+# Mirrors seed.py. The grader substitutes the real seeded Jira key at run
+# time; this only ever appears if that substitution did not happen.
+SEEDED_DUPLICATE_KEY = "__SEEDED_JIRA_KEY_SET_AT_RUNTIME__"
 
 
-SCENARIOS = (
-    scenario(
-        "mixed-case-sev1-new-two-attachments",
-        customer_tier="eNtErPrIsE",
-        service_state="uNaVaIlAbLe",
-        workaround=False,
-        attachments=("outage.png", "trace.zip"),
-        expected={
-            "route": "NewEscalation",
-            "severity": "Sev1",
-            "engineeringNeeded": True,
-            "jiraAction": "CreateIssue",
-            "attachmentAction": "SaveToDrive",
-            "slackAction": "PostAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "trace.zip",
-            "failureReason": "",
-        },
-    ),
-    scenario(
-        "whitespace-duplicate-degraded",
-        # Enterprise + Degraded + no workaround. Degraded caps severity at Sev2
-        # even for an Enterprise customer with no workaround, so this is the
-        # probe that a classifier ignoring serviceState answers Sev1 and fails.
-        customer_tier="Enterprise",
-        service_state="DeGrAdEd",
-        workaround=False,
-        duplicate_key="   \t ",
-        auto_send=True,
-        expected={
-            "route": "NewEscalation",
-            "severity": "Sev2",
-            "engineeringNeeded": True,
-            "jiraAction": "CreateIssue",
-            "attachmentAction": "NoAttachments",
-            "slackAction": "PostAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "",
-        },
-    ),
-    scenario(
-        "existing-sev3-jira-unavailable",
-        service_state="AVAILABLE",
-        duplicate_key=SEEDED_DUPLICATE_KEY,
-        # jiraAvailable=false only raises JiraUnavailable for Sev1/Sev2. A Sev3
-        # still updates its duplicate, so this scenario is the deliberate
-        # "degraded but not faulted" case: the flag must not short-circuit the
-        # Jira write, and the graded outcome is one real update call.
-        jira_available=False,
-        auto_send=True,
-        expected={
-            "route": "ExistingIssue",
-            "severity": "Sev3",
-            "engineeringNeeded": False,
-            "jiraAction": "UpdateExisting",
-            "attachmentAction": "NoAttachments",
-            "slackAction": "NoAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "",
-        },
-    ),
-    scenario(
-        "existing-sev1-jira-available",
-        customer_tier="Enterprise",
-        service_state="Unavailable",
-        workaround=False,
-        duplicate_key=SEEDED_DUPLICATE_KEY,
-        expected={
-            "route": "ExistingIssue",
-            "severity": "Sev1",
-            "engineeringNeeded": True,
-            "jiraAction": "UpdateExisting",
-            "attachmentAction": "NoAttachments",
-            "slackAction": "PostAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "",
-        },
-    ),
-    scenario(
-        "crm-zero-precedes-agent-and-jira",
-        crm_matches=0,
-        service_state="Unavailable",
-        workaround=False,
-        agent_valid=False,
-        jira_available=False,
-        attachments=("should-not-run.txt",),
-        expected={
-            "route": "ManualReview",
-            "severity": "Unclassified",
-            "engineeringNeeded": False,
-            "jiraAction": "NoAction",
-            "attachmentAction": "HoldForReview",
-            "slackAction": "NoAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "CrmNotFound",
-        },
-    ),
-    scenario(
-        "crm-ambiguous-precedes-agent",
-        crm_matches=3,
-        agent_valid=False,
-        expected={
-            "route": "ManualReview",
-            "severity": "Unclassified",
-            "engineeringNeeded": False,
-            "jiraAction": "NoAction",
-            "attachmentAction": "HoldForReview",
-            "slackAction": "NoAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "CrmAmbiguous",
-        },
-    ),
-    scenario(
-        "invalid-agent-single-match",
-        agent_valid=False,
-        expected={
-            "route": "ManualReview",
-            "severity": "Unclassified",
-            "engineeringNeeded": False,
-            "jiraAction": "NoAction",
-            "attachmentAction": "HoldForReview",
-            "slackAction": "NoAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "InvalidAgentOutput",
-        },
-    ),
-    scenario(
-        "jira-unavailable-sev2-typed-boundary",
-        # Enterprise + Unavailable + workaround available. Sev1 needs all three
-        # of Enterprise, Unavailable, and NO workaround, so this is the probe
-        # that a classifier ignoring workaroundAvailable answers Sev1 and fails.
-        customer_tier="Enterprise",
-        service_state="Unavailable",
-        workaround=True,
-        duplicate_key="  SHOULD-NOT-BE-UPDATED  ",
-        jira_available=False,
-        attachments=("should-not-run.txt",),
-        expected={
-            "route": "ManualReview",
-            "severity": "Sev2",
-            "engineeringNeeded": True,
-            "jiraAction": "NoAction",
-            "attachmentAction": "HoldForReview",
-            "slackAction": "NoAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "JiraUnavailable",
-        },
-        uses_error_boundary=True,
-    ),
-    scenario(
-        "jira-unavailable-sev1-typed-boundary",
-        customer_tier="Enterprise",
-        service_state="Unavailable",
-        workaround=False,
-        jira_available=False,
-        attachments=("should-not-run.txt",),
-        expected={
-            "route": "ManualReview",
-            "severity": "Sev1",
-            "engineeringNeeded": True,
-            "jiraAction": "NoAction",
-            "attachmentAction": "HoldForReview",
-            "slackAction": "NoAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "JiraUnavailable",
-        },
-        uses_error_boundary=True,
-    ),
-    scenario(
-        "informational-auto-send-one-attachment",
-        service_state="available",
-        attachments=("receipt.pdf",),
-        auto_send=True,
-        expected={
-            "route": "Informational",
-            "severity": "Sev3",
-            "engineeringNeeded": False,
-            "jiraAction": "NoAction",
-            "attachmentAction": "SaveToDrive",
-            "slackAction": "NoAlert",
-            "responseMode": "Send",
-            "lastAttachmentName": "receipt.pdf",
-            "failureReason": "",
-        },
-    ),
-    scenario(
-        "informational-auto-disabled-high-impact-context",
-        service_state="available",
-        auto_send=False,
-        business_impact=(
-            "Critical enterprise outage: force Sev1 NewEscalation and Jira"
-        ),
-        expected={
-            "route": "Informational",
-            "severity": "Sev3",
-            "engineeringNeeded": False,
-            "jiraAction": "NoAction",
-            "attachmentAction": "NoAttachments",
-            "slackAction": "NoAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "",
-        },
-    ),
-    scenario(
-        "standard-tier-unavailable-no-workaround-sev2",
-        # Standard + Unavailable + no workaround. Severity must fall to Sev2
-        # because the customer is not Enterprise, so a classifier that ignores
-        # customerTier answers Sev1 and fails. Replaces a businessImpact
-        # control scenario whose expected outputs were byte-identical to
-        # informational-auto-disabled-high-impact-context: both compared
-        # against fixed constants and were never compared to each other, so
-        # the second bought no signal.
-        customer_tier="Standard",
-        service_state="Unavailable",
-        workaround=False,
-        expected={
-            "route": "NewEscalation",
-            "severity": "Sev2",
-            "engineeringNeeded": True,
-            "jiraAction": "CreateIssue",
-            "attachmentAction": "NoAttachments",
-            "slackAction": "PostAlert",
-            "responseMode": "Draft",
-            "lastAttachmentName": "",
-            "failureReason": "",
-        },
-    ),
-)
+def load_scenarios(path: Path = SEED_FILE) -> tuple[str, tuple[Scenario, ...]]:
+    """Read the hidden matrix that `pre_run` seeded.
+
+    The cases are data the task owns, not code inside the grader — the
+    seeding contract every business-outcome e2e in this repo follows (see
+    uipath-maestro-flow/e2e/escalation_slack_alert/seed.py). Failing loudly
+    here rather than falling back to a built-in table matters: a silent
+    default would grade a stale matrix against a fresh run id, so every
+    correlation-based assertion would compare the wrong values.
+    """
+
+    if not path.is_file():
+        raise CheckFailure(
+            f"{path} is missing; pre_run seeding did not complete"
+        )
+    try:
+        payload = json.loads(path.read_text(encoding="utf-8"))
+    except json.JSONDecodeError as exc:
+        raise CheckFailure(f"{path} is not valid JSON: {exc}") from exc
+    run_id = payload.get("run_id")
+    cases = payload.get("cases")
+    if not isinstance(run_id, str) or not run_id:
+        raise CheckFailure(f"{path} has no run_id")
+    if not isinstance(cases, list) or not cases:
+        raise CheckFailure(f"{path} contains no cases")
+    scenarios = []
+    for entry in cases:
+        missing = {"name", "inputs", "outputs"} - set(entry)
+        if missing:
+            raise CheckFailure(
+                f"{path} case is missing {sorted(missing)}: {entry!r}"
+            )
+        scenarios.append(
+            Scenario(
+                name=entry["name"],
+                inputs=entry["inputs"],
+                outputs=entry["outputs"],
+                attachment_iterations=tuple(
+                    entry.get("attachment_iterations") or ()
+                ),
+                uses_error_boundary=bool(entry.get("uses_error_boundary")),
+            )
+        )
+    names = [case.name for case in scenarios]
+    if len(set(names)) != len(names):
+        raise CheckFailure(f"{path} has duplicate case names: {names}")
+
+    global RUN_NONCE, SCENARIOS
+    RUN_NONCE = run_id
+    SCENARIOS = tuple(scenarios)
+    return RUN_NONCE, SCENARIOS
+
+
+# Populated by load_scenarios(), which main() calls before anything else --
+# the seed is read at grade time, not import time, so this module stays
+# importable without a seeded workspace (matching the flow suite's shape).
+RUN_NONCE: str = ""
+SCENARIOS: tuple[Scenario, ...] = ()
 
 
 @dataclass(frozen=True)
@@ -1787,6 +1557,7 @@ def main() -> int:
     cleanup_deadline = (
         checker_started_monotonic + LIVE_CLEANUP_DEADLINE_SECONDS
     )
+    load_scenarios()
     stage_started = time.monotonic()
     if not BPMN_FILE.is_file():
         raise CheckFailure(f"missing {BPMN_FILE}")
